+++ lib/easy_ads_flutter.dart
@@ -5,3 +5,5 @@ export 'package:easy_ads_flutter/src/easy_admob/easy_admob_banner_ad.dart';
 export 'package:easy_ads_flutter/src/easy_unity/easy_unity_banner_ad.dart';
 export 'package:easy_ads_flutter/src/easy_ads.dart';
 export 'package:easy_ads_flutter/src/utils/i_ad_id_manager.dart';
+export 'package:easy_ads_flutter/src/easy_ad_base.dart';
+export 'package:google_mobile_ads/google_mobile_ads.dart';